backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
fonttools==4.49.0
gpxpy==1.6.2
idna==3.6
importlib_resources==6.1.2
kiwisolver==1.4.5
matplotlib==3.8.3
mplsoccer==1.2.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.17.0
